Move IStorageEngineMethods to separate file
