+++ arangod/Replication2/ReplicatedState/CMakeLists.txt
@@ -1,5 +1,4 @@
 target_sources(arango_replication2 PRIVATE
-  PersistedStateInfo.cpp
   ReplicatedStateFeature.cpp
   ReplicatedStateMetrics.cpp
   StateCommon.cpp
