+++ arangod/StorageEngine/StorageEngine.cpp
@@ -28,7 +28,7 @@
 
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Replication2/ReplicatedLog/LogCommon.h"
-#include "Replication2/ReplicatedState/PersistedStateInfo.h"
+#include "Replication2/Storage/IStorageEngineMethods.h"
 #include "VocBase/VocbaseInfo.h"
 #include "VocBase/vocbase.h"
