+++ tests/Replication2/ReplicatedLog/Components/FollowerCommitManagerTest.cpp
@@ -27,7 +27,6 @@
 #include "Replication2/ReplicatedLog/Components/IStorageManager.h"
 #include "Replication2/ReplicatedLog/InMemoryLog.h"
 #include "Replication2/ReplicatedLog/TermIndexMapping.h"
-#include "Replication2/ReplicatedState/PersistedStateInfo.h"
 #include "Replication2/ReplicatedLog/ReplicatedLog.h"
 #include "Basics/Result.h"
 #include "Replication2/Mocks/SchedulerMocks.h"
